move timeout into its own package
